requirements.txt
atomicwrites==1.4.0
attrs==21.4.0
click==8.1.0
colorama==0.4.4
Flask==2.1.0
greenlet==1.1.2
importlib-metadata==4.11.3
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
mysql-connector==2.2.9
mysqlclient==2.1.0
numpy==1.22.3
packaging==21.3
pandas==1.4.1
pluggy==1.0.0
psycopg2==2.9.3
py==1.11.0
PyMySQL==1.0.2
pyparsing==3.0.7
pytest==7.1.1
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
SQLAlchemy==1.4.32
tomli==2.0.1
Werkzeug==2.1.0
zipp==3.7.0
